Move enum, type, and interface definitions into /models
