+++ src/app/models/Width.ts
