+++ src/app/routes/Styleguide/index.tsx
@@ -6,8 +6,8 @@ import { Section    } from '~/components/Section'
 import { SideNav    } from '~/components/SideNav'
 import { examples   } from '~/components/utils'
 import * as Box       from '~/components/Box'
-import * as Alignment from '~/components/Box/types/Alignment'
-import * as Width     from '~/components/Box/types/Width'
+import * as Alignment from '~/models/Alignment'
+import * as Width     from '~/models/Width'
 import * as ScrollTo  from '~/components/ScrollTo'
 import * as T         from '~/components/Typography'
